+++ Assets/Scenes/Debug/Vision.unity
@@ -98,7 +98,7 @@ LightmapSettings:
     m_TrainingDataDestination: TrainingData
     m_LightProbeSampleCountMultiplier: 4
   m_LightingDataAsset: {fileID: 0}
-  m_UseShadowmask: 0
+  m_UseShadowmask: 1
 --- !u!196 &4
 NavMeshSettings:
   serializedVersion: 2
@@ -198,13 +198,13 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 76384031}
-  m_LocalRotation: {x: 0, y: -1, z: 0, w: 0}
-  m_LocalPosition: {x: 7, y: 1, z: 41}
+  m_LocalRotation: {x: -0, y: -0.95611596, z: 0.2929885, w: 0}
+  m_LocalPosition: {x: 9.92, y: 5.84, z: 37.07}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: -180, z: 0}
+  m_LocalEulerAnglesHint: {x: 34.074, y: -180.00002, z: 0}
 --- !u!114 &76384035
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -382,13 +382,13 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 134363184}
-  m_LocalRotation: {x: -0.07550301, y: -0.44792995, z: 0.79322374, w: 0.4055295}
+  m_LocalRotation: {x: 0.044890847, y: -0.6580183, z: 0.62996596, w: 0.41004828}
   m_LocalPosition: {x: 0, y: 3, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 1
-  m_LocalEulerAnglesHint: {x: 40.495003, y: -39.438004, z: 110.781006}
+  m_LocalEulerAnglesHint: {x: 59.983, y: -74.939, z: 66.153}
 --- !u!1001 &302539353
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1358,6 +1358,7 @@ GameObject:
   - component: {fileID: 1915472592}
   - component: {fileID: 1915472596}
   - component: {fileID: 1915472591}
+  - component: {fileID: 1915472597}
   m_Layer: 0
   m_Name: UV Cube
   m_TagString: Untagged
@@ -1467,6 +1468,23 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e72e8434866d495bb200b94a27b19801, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1915472597
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1915472590}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7ae2607a5bc2dcb48affc96cbc1db46e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  MaterialGroups:
+  - originalMaterial: {fileID: 2100000, guid: 1d252bf4d5f4544d39cab982e9ffe6d1, type: 2}
+    MaterialsToSwap:
+    - animal: 1
+      material: {fileID: 2100000, guid: f91283e6f40144071a73cecf27444b62, type: 2}
 --- !u!1 &2068513153
 GameObject:
   m_ObjectHideFlags: 0
